+++ TunerPage.py
@@ -44,10 +44,9 @@ class Tuner(threading.Thread):
         self.bouton_stop.pack()
         self.filename="output.wav" #nom du fichier temporaire ou est enregistré le son
         self.p=p
+        self.rm=0
 
     def run(self):
-        if os.path.exists(self.filename):
-                    os.remove(self.filename) #supprime le fichier temporaire
         tab = gen_frequences() #création du tableau des fréquences pour détection de la note
 
         sample_format = pyaudio.paInt16  # 16 bits per sample
@@ -59,10 +58,12 @@ class Tuner(threading.Thread):
 
 
         while not self._stopevent.isSet():
-            while self.enpause:
-                if os.path.exists(self.filename):
+            while self.enpause :
+                if os.path.exists(self.filename) and self.rm==1:
                     os.remove(self.filename) #supprime le fichier temporaire
+                    self.rm=0
                 time.sleep(0.5)
+                
                 
             #Ouverture du micro
             stream = self.p.open(format=sample_format,
@@ -102,8 +103,6 @@ class Tuner(threading.Thread):
 
             freq_in_hertz = abs(freq * frate)
             res=find_note(tab,freq_in_hertz)
-            # tab_MIDI_song.append(find_Midi_Note(res[0]))
-            #print("code MIDI de ",res,": ",find_Midi_Note(res[0])) #TODO à supprimer : affichage du code midi pour chaque note entendue, ça marche bien
             notePrec = find_note(tab, freq_in_hertz/coeff_frequences) #nom de la note précédente #TODO ? juste limiter à notePrec[0] pour avoir que le nom (le reste sert à rien pour cette utilisation)
             noteSuiv = find_note(tab, freq_in_hertz*coeff_frequences) #nom de la note suivante
             if(res[0]!='-'):
@@ -131,30 +130,7 @@ class Tuner(threading.Thread):
 
     def pause(self):
         self.enpause=True
-        # # create your MIDI object
-        # mf = MIDIFile(1)     # only 1 track
-        # track = 0   # the only track
-
-        # time = 0    # start at the beginning
-        # mf.addTrackName(track, time, "Sample Track")
-        # mf.addTempo(track, time, 200)
-
-        # # add some notes
-        # channel = 0
-        # volume = 100
-
-        # for i in range(len(tab_MIDI_song)):
-        #     if(tab_MIDI_song[i] == "-"): #si pas de note on laisse un silence
-        #         time+=1
-        #     else: #sinon on ajoute la note avec une durée de 1
-        #         pitch = int(tab_MIDI_song[i]) 
-        #         time+=1
-        #         duration = 1
-        #         mf.addNote(track, channel, pitch, time, duration, volume)
-
-        # # write it to disk
-        # with open("output.mid", 'wb') as outf: #on écrit dans le fichier MIDI
-        #     mf.writeFile(outf)
+        self.rm=1
         
 
     def lancer(self):
